Keep the cut inside the maintainability limits
The two preceding commits each pushed one function or file past a
pkg-maint limit. Three fixes, no behavior change:

- RootErrorResponse went to complexity 16 because inlining errors.As added a
  second condition. The classification lookup moves to
  classifiedInferenceErrorResponse next to the other inference-failure
  mapping, which reads better than a bare errors.As in the middle of sentinel
  matching.
- provideModelsCLIInvocationOperation and its adapter move from profiles.go
  (1003 lines, limit 1000) to models_runtime.go, which already owns the
  Models wire providers and already imports most of what the adapter needs.
- TestGetFactorySessionResults_ProjectsFailureAndPrimaryResult reached
  complexity 21 by asserting two separate claims. It splits into an identity
  and echo test and a failure-detail and payload test over a shared request
  helper, so a failure now names which half regressed.

Also formats the two struct literal lines the equivalence-coverage commit
left unaligned.
