+++ pkg/services/models/transports/http/inference_failure_mapping.go
@@ -1,9 +1,11 @@
 package http
 
 import (
+	"errors"
 	"net/http"
 
 	"github.com/portpowered/infinite-you/pkg/services/models"
+	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 )
 
 func inferenceFailureHTTPStatus(failure *models.InferenceFailure) int {
@@ -34,4 +36,15 @@ func inferenceFailureErrorCode(failure *models.InferenceFailure) string {
 	default:
 		return "MODEL_INFERENCE_RUNTIME_FAILURE"
 	}
+}
+
+// classifiedInferenceErrorResponse maps an already-classified inference failure
+// anywhere in err's chain to its public response. It reports false when err
+// carries no classification, leaving the caller's sentinel matching to decide.
+func classifiedInferenceErrorResponse(err error) (int, factoryapi.ErrorResponse, bool) {
+	var failure *models.InferenceFailure
+	if !errors.As(err, &failure) || failure == nil {
+		return 0, factoryapi.ErrorResponse{}, false
+	}
+	return inferenceFailureErrorResponse(failure)
 }
