+++ pkg/wire/models_runtime.go
@@ -17,8 +17,10 @@ import (
 	platformprocess "github.com/portpowered/infinite-you/pkg/platform/process"
 	platformrandom "github.com/portpowered/infinite-you/pkg/platform/random"
 	serviceedges "github.com/portpowered/infinite-you/pkg/services/edges"
+	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
 	factorysessionwire "github.com/portpowered/infinite-you/pkg/services/factory_sessions/wire"
 	"github.com/portpowered/infinite-you/pkg/services/models"
+	modelscli "github.com/portpowered/infinite-you/pkg/services/models/transports/cli"
 	modelswire "github.com/portpowered/infinite-you/pkg/services/models/wire"
 	"github.com/portpowered/infinite-you/pkg/services/workers"
 	workerswire "github.com/portpowered/infinite-you/pkg/services/workers/wire"
@@ -417,4 +419,48 @@ func (p *modelsManagedProcess) Stop(ctx context.Context) error {
 	case <-ctx.Done():
 		return ctx.Err()
 	}
+}
+
+func provideModelsCLIInvocationOperation(
+	invocation factorysessionwire.InvocationOperation,
+) modelscli.InvocationOperation {
+	if invocation == nil {
+		return nil
+	}
+	return modelsCLIInvocationOperation{invocation: invocation}
+}
+
+// modelsCLIInvocationOperation maps the four invocation inputs the Models CLI
+// resolves onto the fuller Factory Sessions invocation target. Owning the
+// mapping here keeps the Models CLI transport free of a Factory Sessions
+// import while preserving the exact values the command sent before.
+type modelsCLIInvocationOperation struct {
+	invocation factorysessionwire.InvocationOperation
+}
+
+func (o modelsCLIInvocationOperation) InvokeModel(
+	ctx context.Context,
+	target modelscli.InvocationTarget,
+	modelName string,
+	request models.Request,
+) (models.Result, error) {
+	return o.invocation.InvokeModel(ctx, factorysessions.InvocationTarget{
+		FactoryDir:       target.FactoryDir,
+		HomeDir:          target.HomeDir,
+		OperatorDefaults: target.OperatorDefaults,
+		Verbose:          target.Verbose,
+	}, modelName, request)
+}
+
+func (o modelsCLIInvocationOperation) ResolveModelInvocationFactoryDir(
+	factoryDir string,
+) (string, error) {
+	return o.invocation.ResolveModelInvocationFactoryDir(factoryDir)
+}
+
+func (o modelsCLIInvocationOperation) ExportModelInvocationArtifact(
+	source string,
+	destination string,
+) error {
+	return o.invocation.ExportModelInvocationArtifact(source, destination)
 }
